[feat] BattleSystem
-feat BattleSystem 정적클래스 수정
-feat Manager 던전 매니저에서 전역 처리
-fix BattleSystem 전투 종료 시 층 수 증가 수정
