+++ ConsoleTextRPG/Manager.cs
@@ -1,4 +1,6 @@
-﻿using GameLogic;
+﻿using BattleSystem;
+using System.Drawing;
+using GameLogic;
 using GameQuest;
 using GameService;
 
@@ -30,6 +32,31 @@ namespace Manager
             player = _playerJob;
             player.SetName(_playerName);
             quest.Load(false);
+        }
+        
+        public static int DungeonFloor { get; set; } = 1;
+        public static void NextFloor()
+        {
+            DungeonFloor++;
+            if (DungeonFloor == 10)
+            {
+                Console.WriteLine("던전 클리어!");
+                Console.WriteLine("축하합니다!");
+                Console.WriteLine("무한모드로 진입합니다.");
+                Console.ReadKey();
+            }
+            else if (DungeonFloor > 10)
+            {
+                Console.WriteLine("무한모드입니다.");
+                Console.ReadKey();
+            }
+            else
+            {
+                Console.WriteLine($"던전 {DungeonFloor}층으로 이동합니다...");
+                Thread.Sleep(1000);
+                Console.Clear();
+                BattleSystems.Start();
+            }
         }
     }
 }
